move clear button to text input
