+++ components/Input/TextInput.tsx
@@ -1,10 +1,12 @@
 import { useHandleDecimalInput } from "hooks";
 import styled from "styled-components";
 import { Input, Wrapper } from "./Input";
+import Close from "public/assets/icons/x.svg";
 
 interface Props {
   value: string;
   onInput: (value: string) => void;
+  onClear?: () => void;
   disabled?: boolean;
   placeholder?: string;
   type?: "text" | "number";
@@ -15,6 +17,7 @@ interface Props {
 export function TextInput({
   value,
   onInput,
+  onClear,
   disabled,
   placeholder,
   type = "text",
@@ -46,11 +49,17 @@ export function TextInput({
         maxLength={79}
         spellCheck="false"
       />
+      {!!onClear && (
+        <ClearInputButton aria-label="exit custom input" onClick={onClear}>
+          <CloseIcon />
+        </ClearInputButton>
+      )}
     </_Wrapper>
   );
 }
 
 const _Wrapper = styled(Wrapper)`
+  position: relative;
   &[aria-disabled="true"] {
     opacity: 0.25;
   }
@@ -59,3 +68,22 @@ const _Wrapper = styled(Wrapper)`
 const _Input = styled(Input)`
   padding-left: 15px;
 `;
+
+const ClearInputButton = styled.button`
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  display: grid;
+  place-items: center;
+  width: 24px;
+  height: 24px;
+  border-radius: 50%;
+  background: var(--grey-100);
+  transition: opacity 0.2s;
+
+  &:hover {
+    opacity: 0.75;
+  }
+`;
+
+const CloseIcon = styled(Close)``;
